Project/requirements.txt
flask
cohere
python-dotenv